칼럼변경에 따른 변경

+++ src/main/java/com/cook/mymealkit/service/CartService.java
@@ -6,9 +6,10 @@ import com.cook.mymealkit.domain.CartDTO;
 
 public interface CartService {
 	
-	public void cartInsert(CartDTO dto);
-	public List<CartDTO> cartList();
-	public CartDTO cartFindById(int cart_id);
-	public void cartUpdate(CartDTO dto);
-	public void cartDelete(int cart_id); 
+	public void cartInsert(CartDTO dto);			// 등록
+	public List<CartDTO> cartList();				// 목록
+	public CartDTO cartFindById(long cart_id);		// 조회
+	public void cartUpdate(CartDTO dto);			// 수정
+	public void cartDelete(long cart_id); 			// 삭제
+	
 }
